--- a/app/src/androidTest/java/com/duckduckgo/app/statistics/VariantManagerTest.kt
+++ b/app/src/androidTest/java/com/duckduckgo/app/statistics/VariantManagerTest.kt
@@ -43,28 +43,31 @@ class VariantManagerTest {
         assertEquals(0, variant.features.size)
     }
 
-    // Single Search Bar Experiments
+    // Use our app experiment
     @Test
-    fun serpHeaderControlVariantHasExpectedWeightAndNoFeatures() {
-        val variant = variants.first { it.key == "zg" }
+    fun inBrowserControlVariantHasExpectedWeightAndNoFeatures() {
+        val variant = variants.first { it.key == "zj" }
         assertEqualsDouble(0.0, variant.weight)
         assertEquals(0, variant.features.size)
     }
 
     @Test
-    fun serpHeaderVariantHasExpectedWeightAndSERPHeaderRemovalFeature() {
-        val variant = variants.first { it.key == "zi" }
+    fun inBrowserSecondControlVariantHasExpectedWeightAndRemoveDay1And3NotificationsAndKillOnboardingFeatures() {
+        val variant = variants.first { it.key == "zk" }
         assertEqualsDouble(0.0, variant.weight)
-        assertEquals(1, variant.features.size)
-        assertEquals(SerpHeaderRemoval, variant.features[0])
+        assertEquals(2, variant.features.size)
+        assertTrue(variant.hasFeature(KillOnboarding))
+        assertTrue(variant.hasFeature(RemoveDay1AndDay3Notifications))
     }
 
     @Test
-    fun serpHeaderVariantHasExpectedWeightAndSERPHeaderQueryReplacementFeature() {
-        val variant = variants.first { it.key == "zh" }
+    fun inBrowserInAppUsageVariantHasExpectedWeightAndRemoveDay1And3NotificationsAndKillOnboardingAndInAppUsageFeatures() {
+        val variant = variants.first { it.key == "zl" }
         assertEqualsDouble(0.0, variant.weight)
-        assertEquals(1, variant.features.size)
-        assertEquals(SerpHeaderQueryReplacement, variant.features[0])
+        assertEquals(3, variant.features.size)
+        assertTrue(variant.hasFeature(KillOnboarding))
+        assertTrue(variant.hasFeature(RemoveDay1AndDay3Notifications))
+        assertTrue(variant.hasFeature(InAppUsage))
     }
 
     @Test