--- a/app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
+++ b/app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
@@ -79,6 +79,7 @@ import com.duckduckgo.app.global.plugins.view_model.ViewModelFactoryPlugin
 import com.duckduckgo.app.global.useourapp.UseOurAppDetector
 import com.duckduckgo.app.global.useourapp.UseOurAppDetector.Companion.USE_OUR_APP_SHORTCUT_TITLE
 import com.duckduckgo.app.global.useourapp.UseOurAppDetector.Companion.USE_OUR_APP_SHORTCUT_URL
+import com.duckduckgo.app.global.view.DaxDialog
 import com.duckduckgo.app.global.view.asLocationPermissionOrigin
 import com.duckduckgo.app.globalprivacycontrol.GlobalPrivacyControl
 import com.duckduckgo.app.location.GeoLocationPermissions
@@ -1619,7 +1620,7 @@ class BrowserTabViewModel(
     }
 
     suspend fun refreshCta(locale: Locale = Locale.getDefault()): Cta? {
-        if (currentGlobalLayoutState() is Browser) {
+        if (currentGlobalLayoutState() is Browser && currentCtaViewState().cta !is DaxDialog) {
             val cta = withContext(dispatchers.io()) {
                 ctaViewModel.refreshCta(dispatchers.io(), currentBrowserViewState().browserShowing, siteLiveData.value, locale)
             }