--- a/app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
+++ b/app/src/main/java/com/duckduckgo/app/browser/BrowserTabViewModel.kt
@@ -1612,12 +1612,8 @@ class BrowserTabViewModel(
         ctaViewModel.onCtaShown(cta)
     }
 
-    fun onManualCtaShown(cta: Cta) {
-        ctaViewModel.onCtaShown(cta)
-    }
-
     suspend fun refreshCta(locale: Locale = Locale.getDefault()): Cta? {
-        if (currentGlobalLayoutState() is Browser && currentCtaViewState().cta !is DaxDialog) {
+        if (currentGlobalLayoutState() is Browser) {
             val cta = withContext(dispatchers.io()) {
                 ctaViewModel.refreshCta(dispatchers.io(), currentBrowserViewState().browserShowing, siteLiveData.value, locale)
             }