--- a/app/src/main/java/com/duckduckgo/app/browser/shortcut/ShortcutReceiver.kt
+++ b/app/src/main/java/com/duckduckgo/app/browser/shortcut/ShortcutReceiver.kt
@@ -24,14 +24,23 @@ import android.widget.Toast
 import com.duckduckgo.app.browser.R
 import com.duckduckgo.app.browser.shortcut.ShortcutBuilder.Companion.SHORTCUT_TITLE_ARG
 import com.duckduckgo.app.browser.shortcut.ShortcutBuilder.Companion.SHORTCUT_URL_ARG
+import com.duckduckgo.app.global.DispatcherProvider
+import com.duckduckgo.app.global.events.db.UserEventKey
+import com.duckduckgo.app.global.events.db.UserEventsStore
 import com.duckduckgo.app.global.useourapp.UseOurAppDetector
 import com.duckduckgo.app.pixels.AppPixelName
+import com.duckduckgo.app.statistics.VariantManager
 import com.duckduckgo.app.statistics.pixels.Pixel
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class ShortcutReceiver @Inject constructor(
     private val useOurAppDetector: UseOurAppDetector,
-    private val pixel: Pixel
+    private val pixel: Pixel,
+    private val userEventsStore: UserEventsStore,
+    private val dispatcher: DispatcherProvider,
+    private val variantManager: VariantManager
 ) :
     BroadcastReceiver() {
 
@@ -45,10 +54,15 @@ class ShortcutReceiver @Inject constructor(
             }
         }
 
-        if (useOurAppDetector.isUseOurAppUrl(originUrl)) {
-            pixel.fire(AppPixelName.USE_OUR_APP_SHORTCUT_ADDED)
-        } else {
-            pixel.fire(AppPixelName.SHORTCUT_ADDED)
+        GlobalScope.launch(dispatcher.io()) {
+            if (useOurAppDetector.isUseOurAppUrl(originUrl)) {
+                pixel.fire(AppPixelName.USE_OUR_APP_SHORTCUT_ADDED)
+                if (variantManager.getVariant().hasFeature(VariantManager.VariantFeature.InAppUsage)) {
+                    userEventsStore.registerUserEvent(UserEventKey.USE_OUR_APP_SHORTCUT_ADDED)
+                }
+            } else {
+                pixel.fire(AppPixelName.SHORTCUT_ADDED)
+            }
         }
     }