… app/src/main/java/com/duckduckgo/app/di/NotificationModule.kt
… app/src/main/java/com/duckduckgo/app/global/plugins/worker/WorkerPluginsModule.kt
… app/src/main/java/com/duckduckgo/app/job/JobCleaner.kt
… app/src/main/java/com/duckduckgo/app/notification/AndroidNotificationScheduler.kt
… app/src/main/java/com/duckduckgo/app/onboarding/store/UserStageStore.kt
… statistics/src/main/java/com/duckduckgo/app/statistics/VariantManager.kt