<!--
Note: This checklist is a reminder of our shared engineering expectations. Feel free to change it, although assigning a GitHub reviewer and the items in bold are required.
-->

Task/Issue URL: https://app.asana.com/0/1125189844152671/1200123506690912
Tech Design URL: 
CC: 

**Description**:
This PR re-adds the code removed in https://github.com/duckduckgo/Android/pull/924 to re-enable the use our app experiment.

**Before testing**
In `AppUserStageStore` lines 83 and 84 change the code to:

```
val days = TimeUnit.MILLISECONDS.toSeconds(System.currentTimeMillis() - appInstallStore.installTimestamp)
if (days >= 30) {
```

You will be now moved out of the onboarding stage after 30 seconds from install whenever you resume the app.

In `AndroidNotificationScheduler` lines 77 and 89 change the code to be `40` and `TimeUnit.SECONDS`. Also change the line 58 to be:

```
scheduleNotification(OneTimeWorkRequestBuilder<PrivacyNotificationWorker>(), 40, TimeUnit.SECONDS, UNUSED_APP_WORK_REQUEST_TAG)
```

The notifications will be now triggered after 40 seconds.

Go to `CtaViewModel` lines 259 & 260 and change them to:

```
val days = TimeUnit.MILLISECONDS.toMinutes(System.currentTimeMillis() - timestampKey.timestamp)
return (days >= 1)
```

This will now trigger the second CTA after 1 minute rather than 2 days.

### **Steps to test this PR**:

#### **Variant mc - In App usage flow**
Hardcode the variant to use `mc`

**Normal flow**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. The new in-app usage notification should show up.
1. Pixel `mnot_s_uoa` should be sent (notification seen)
1. Tap the notification.
1. Pixel `mnot_l_uoa` should be sent (notification launched)
1. Home screen should show, the keyboard should be hidden and a the new cta should shown.
1. Tap on "Add Facebook Shortcut".
1. Pixel `m_uoa_d_ok` should be sent (use our app ok button pressed)
1. Pixel `m_uoa_van` should be sent (use our app site visited after notification seen)
1. Add the shortcut
1. Pixel `m_uoa_s_a` should be sent. (use our app shortcut added)
1. A toast message saying the shortcut was added should show (message can vary depending on the device).
1. Sign in to Facebook
1. The fireproof dialog should be shown
1. Close the facebook tab and navigate to facebook again.
1. Pixel `m_uoa_vas` should be sent (use our app site visited after adding shortcut).
1. After a minute (or whatever time you set in `CtaViewModel` the next time you navigate to Facebook (page has to load) you should see a new Cta.
1. Pixel `m_uoa_dd` should be sent (use our app delete dialog seen)
1. Close the facebook tab and open facebook again in a new tab.
1. Pixel `m_uoa_vad` should be sent (use our app site visited after delete cta).

**Clicking the notification moves you out of the onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. The new in-app usage notification should show up.
1. Notice the day 1 privacy notification does not show up.
1. Click on the notification, you should be taken to a new tab showing the hombre screen with the new CTA.
1. Ignore the CTA (click Not now) and go to `instagram.com`.
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**After X time you are moved out of the onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 30 seconds (if you see the in app usage notification do not interact with it or change the timing to a higher value so you don't see it).
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**Hiding tips does not show the notification**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, click hide tips and hide them forever.
1. Stay in that screen for at least 40 seconds 
1. You should not see any notifications.

#### **Variant mb - Removes onboarding and day1/3 notifications**
Hardcode the variant to use `mb`

**After X time you are moved out of the onboarding and no notifications are scheduled**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. No notifications should be shown.
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should not see any onboarding CTAs since you are not in the onboarding stage anymore.

**Pixels and tabs**
1. Launch the app .
1. Go to facebook.com
1. Pixel `m_uoa_v` should be sent.
1. In the menu click Add to Home.
1. The Add to Home Screen dialog should show with a high-res icon.
1. Click in Add.
1. Pixel `m_uoa_s_a` should be sent
1. Login to facebook.
1. The fireproof dialog should be shown or not depending on your settings.
1. On Facebook, navigate to your notifications.
1. Kill the app.
1. Open the shortcut you just created.
1. Pixels `m_sho_uoa_o` and `m_uoa_v` should be sent.
1. The notifications tab you had opened before should be opened instead of a new tab.
1. Close the tab and send the app to the background.
1. Open the shortcut.
1. A new tab should be created pointing to facebook.com

#### **Variant ma - Control**
Hardcode the variant to use `ma`

**Notification and onboarding**
1. Install and launch the app.
1. Go to SERP.
1. You should see the SERP CTA, press Phew.
1. Stay in that screen for at least 40 seconds.
1. The privacy notification should be shown.
1. Send the app to the background or kill the app.
1. Relaunch the app
1. Go to instagram.com
1. You should see a CTA.

**Pixels and tabs**
1. Launch the app .
1. Go to facebook.com
1. Pixel `m_uoa_v` should be sent.
1. In the menu click Add to Home.
1. The Add to Home Screen dialog should show with a high-res icon.
1. Click in Add.
1. Pixel `m_uoa_s_a` should be sent
1. Login to facebook.
1. The fireproof dialog should be shown or not depending on your settings.
1. On Facebook, navigate to your notifications.
1. Kill the app.
1. Open the shortcut you just created.
1. Pixels `m_sho_uoa_o` and `m_uoa_v` should be sent.
1. The notifications tab you had opened before should be opened instead of a new tab.
1. Close the tab and send the app to the background.
1. Open the shortcut.
1. A new tab should be created pointing to facebook.com

---
###### Internal references:
[Software Engineering Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design Template](https://app.asana.com/0/59792373528535/184709971311943)