@marcosholgado 2021-04-07
Hey @cmonfortep this is ready to re-review. We decided to remove the shown pixel for the fist CTA because it was causing too many issues. For the second CTA I noticed we were triggering the pixel way too early, should be fine now.